requirements.txt
-i https://pypi.org/simple
antlr4-python3-runtime==4.9.3
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
dateparser==1.2.0
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
ply==3.11
plyara==2.1.1
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
requests==2.31.0; python_version >= '3.7'
simplejson==3.19.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.1; python_version >= '3.7'
stix2==3.0.1; python_version >= '3.6'
stix2-patterns==2.0.0; python_version >= '3.6'
tqdm==4.66.2; python_version >= '3.7'
urllib3==2.2.1; python_version >= '3.8'
